[mtouch/mmp] Unify OptimizeGeneratedCodeSubStep into a single class.
This makes it easier to use this code from dotnet-linker.
